# add early skips in checkHitspLS