```
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       TCE      Event      Short             Rate
   old      2.7      3.9      3.4      3.4      3.7      2.6      6.5      1.8      3.7      5.8      37.4      32.1+/-  5.3      40.9   explicit_cache[s=1]
   new      2.8      4.0      3.6      3.5      3.7      1.3      5.8      1.8      3.6      5.8      36.0      31.9+/-  5.5      39.8   explicit_cache[s=1]
```
old is 21cf443 new is a5876c4

checkHitspLS average time changes from 1.65ms to 0.96ms (recall that it's average out of 2 calls), so, the total should be down by 1.4ms, as observed:
- the first commit takes it down to 0.99ms
- the second commit to 0.96ms

admittedly, most of the gain is in pLS, which we sort of expect to probably not need a cleaning pass if patatracks come cleaned already