@slava77 2022-05-15
I shaved off a few more computations in `checkHitspLS`

The updated improvement is
```
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       TCE      Event      Short             Rate
   old      2.7      3.9      3.4      3.4      3.7      2.6      6.5      1.8      3.7      5.8      37.4      32.1+/-  5.3      40.9   explicit_cache[s=1]
   new      2.7      3.9      3.5      3.5      3.7      1.1      6.5      1.8      3.4      5.7      35.9      32.0+/-  5.6      39.0  explicit_cache[s=1]
```
where old is 21cf443ae2b6351ecf5c9d473e6ac9960d39d121 new is 5170729

`checkHitspLS` is down from 1.65ms to 0.81ms (1.7ms speedup in total)
- extending parallelization in the outer loop from 779654f: 0.96ms -> 0.88ms
- checking dEta earlier from 5170729: 0.88ms -> 0.81ms