@tresreid 2022-05-16
https://www.classe.cornell.edu/~mgr85/www/SegmentLinking_keep/PR163/eff_plots__GPU_explicit_cache_894a9a8_PU200/#1.5